pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-agent-mail"
version = "0.1.0"
description = "Coordinated multi-agent messaging and coordination MCP server."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "mcp-agent-mail maintainers" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Utilities",
  "Typing :: Typed",
]
dependencies = [
  "fastmcp>=2.10.5",
  "uvicorn[standard]>=0.32.0",
  "fastapi>=0.110.0",
  "starlette>=0.37.2",
  "sse-starlette>=2.2.1",
  "jinja2>=3.1.4",
  "markdown2>=2.4.12",
  "markdown-it-py>=3.0.0",
  "MarkupSafe>=3.0.2",
  "numpy>=2.1.2",
  "bleach>=6.1.0",
  "python-decouple>=3.8",
  "pydantic-settings>=2.6.0",
  "sqlmodel>=0.0.16",
  "sqlalchemy[asyncio]>=2.0.35",
  "aiosqlite>=0.20.0",
  "httpx[http2]>=0.27.2",
  "httpx-sse>=0.4.0",
  "httpcore>=1.0.5",
  "aiolimiter>=1.1.0",
  "anyio>=4.4.0",
  "idna>=3.7",
  "certifi>=2024.8.30",
  "exceptiongroup>=1.2.2",
  "click>=8.1.7",
  "tenacity>=8.3.0",
  "redis[hiredis]>=5.3.0",
  "rich>=13.9.1",
  "sniffio>=1.3.1",
  "typer>=0.15.0",
  "typing-extensions>=4.12.2",
  "structlog>=24.1.0",
  "diskcache>=5.6.3",
  "orjson>=3.10.7,<4",
  "tiktoken>=0.7.0",
  "pillow>=10.4.0",
  "GitPython>=3.1.43",
  "filelock>=3.16.0",
  "py-key-value-aio[disk,keyring,memory]>=0.2.8,<0.3.0",
  "jsonschema>=4.23.0",
  "jsonschema-specifications>=2023.12.1",
  "referencing>=0.35.0",
  "attrs>=24.2.0",
  "pywin32>=306; sys_platform == 'win32'",
  "authlib>=1.5.2,<1.6",
  "cryptography>=43.0.1",
  "litellm>=1.40.0,<2",
  "llama-cpp-python>=0.2.87",
  "pytest>=8.4.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.14.1",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.7",
  "mypy>=1.11.1",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "pytest-xdist>=3.6.1",
  "ipython>=8.27.0",
]

[project.urls]
Homepage = "https://example.com/mcp-agent-mail"
Repository = "https://example.com/mcp-agent-mail"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_agent_mail"]

[tool.hatch.build.targets.sdist]
exclude = ["/.venv", "/.vscode", "/.idea", "/.git", "/__pycache__", "*.py[cod]", "*.db*", ".DS_Store"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "W",
  "F",
  "B",
  "I",
  "C4",
  "ASYNC",
  "A",
  "RUF",
  "SIM",
  "PTH",
  "FA",
]
ignore = [
  "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["mcp_agent_mail"]
combine-as-imports = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
plugins = []
exclude = [
  "^notes/scripts/",
  "^legacy/",
  "^scripts/test_api_smoke\\.py",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mcp_agent_mail.storage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "tests.http_transport_mocks"
ignore_errors = true

[[tool.mypy.overrides]]
module = "scripts.ui_audit_run"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
  "mcp_agent_mail.http",
  "mcp_agent_mail.db",
  "mcp_agent_mail.routers.missions",
  "mcp_agent_mail.workflow_engine",
  "mcp_agent_mail.__init__",
  "agent_missions_hub.db",
  "agent_missions_hub.http",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
  "decouple",
  "decouple.*",
  "requests",
  "requests.*",
  "authlib",
  "authlib.*",
  "playwright",
  "playwright.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
  "--strict-markers",
  "--cov=src/mcp_agent_mail",
  "--cov-config=.coveragerc",
  "--cov-report=term-missing",
]
pythonpath = ["src"]
tmp_path_retention_count = 0
timeout = 60
timeout_method = "thread"
cache_dir = ".pytest_cache_local"
filterwarnings = [
  "ignore:coroutine 'FastMCP.get_tool' was never awaited:RuntimeWarning",
]
markers = [
  "asyncio: mark test as requiring asyncio event loop",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
]





